+++ extendr-api/src/error.rs
@@ -81,6 +81,9 @@ pub enum Error {
 
     ExpectedExternalPtrType(Robj, String),
     Other(String),
+
+    #[cfg(feature = "ndarray")]
+    NDArrayShapeError(ndarray::ShapeError),
 }
 
 impl std::fmt::Display for Error {
@@ -159,6 +162,11 @@ impl std::fmt::Display for Error {
             }
             Error::NoGraphicsDevices(_robj) => write!(f, "No graphics devices active."),
             Error::Other(str) => write!(f, "{}", str),
+
+            #[cfg(feature = "ndarray")]
+            Error::NDArrayShapeError(shape_error) => {
+                write!(f, "NDArray failed with error: {}.", shape_error)
+            }
         }
     }
 }