+++ extendr-api/src/scalar/macros.rs
@@ -791,24 +791,23 @@ macro_rules! gen_trait_impl {
 
 /// Generates an implementation of `std::iter::Sum` for a scalar type
 ///
-/// This macro requires the following arguments:
+/// This macro requires the following argument:
 ///
 /// * `$type`   - The Type to implement `std::iter::Sum` for
-/// * `$zero`   - The zero value for the primitive counterpart to Type
 ///
 /// Example Usage:
 ///
 /// ```ignore
-/// gen_sum_iter!(Rint, 0i32);
+/// gen_sum_iter!(Rint);
 /// ```
 macro_rules! gen_sum_iter {
-    ($type : tt, $zero : expr) => {
+    ($type : tt) => {
         // The 'example usage' expands to...
         //
         // impl std::iter::Sum for $type {
         //     /// Documentation comments/test built by the #[doc] attributes
         //     fn sum<I: Iterator<Item = Rint>>(iter: I) -> Rint {
-        //         iter.fold(Rint::from(0i32), |a, b| a + b)
+        //         iter.fold(Rint::default(), |a, b| a + b)
         //     }
         // }
         impl std::iter::Sum for $type {
@@ -823,7 +822,33 @@ macro_rules! gen_sum_iter {
                 #[doc = "}"]
                 #[doc = "```"]
                 fn sum<I: Iterator<Item = $type>>(iter: I) -> $type {
-                    iter.fold($type::from($zero), |a, b| a + b)
+                    iter.fold($type::default(), |a, b| a + b)
+                }
+            }
+        }
+
+        // The 'example usage' expands to...
+        //
+        // impl std::iter::Sum for &$type {
+        //     /// Documentation comments/test built by the #[doc] attributes
+        //     fn sum<I: Iterator<Item = &Rint>>(iter: I) -> Rint {
+        //         iter.fold(Rint::default(), |a, b| a + b)
+        //     }
+        // }
+        impl<'a> std::iter::Sum<&'a $type> for $type {
+            paste::paste! {
+                #[doc = "Yields NA on overflow if NAs present."]
+                #[doc = "```"]
+                #[doc = "use extendr_api::prelude::*;"]
+                #[doc = "use std::iter::Sum;"]
+                #[doc = "test! {"]
+                #[doc = "    let z =" $type "::default();"]
+                #[doc = "    let x = (0..100).map(|_| &z);"]
+                #[doc = "    assert_eq!(<" $type " as Sum<& " $type ">>::sum(x), <" $type ">::default());"]
+                #[doc = "}"]
+                #[doc = "```"]
+                fn sum<I: Iterator<Item = &'a $type>>(iter: I) -> $type {
+                    iter.fold($type::default(), |a, b| a + *b)
                 }
             }
         }