+++ tests/extendrtests/src/rust/src/lib.rs
@@ -3,6 +3,9 @@ use extendr_api::{graphics::*, prelude::*};
 mod submodule;
 use submodule::*;
 
+mod ndarray;
+use ndarray::*;
+
 mod graphic_device;
 
 // Return string `"Hello world!"` to R.
@@ -199,6 +202,11 @@ fn get_default_value(#[default = "42"] x: i32) -> i32 {
     x
 }
 
+#[extendr(use_try_from = true)]
+fn add_5_if_not_null(x: Nullable<Rint>) -> Nullable<Rint> {
+    x.map(|y| y + 5)
+}
+
 // Class for testing
 #[derive(Default, Debug)]
 struct MyClass {
@@ -338,6 +346,8 @@ extendr_module! {
 
     fn get_default_value;
 
+    fn add_5_if_not_null;
+
     impl MyClass;
     impl __MyClass;
     impl MyClassUnexported;
@@ -345,4 +355,5 @@ extendr_module! {
     fn my_device;
 
     use submodule;
+    use ndarray;
 }